+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 2022-04-28
+### Added
+- Custom 404 page
+- Custom error page
+- Added description, external link to blog article page
+### Changed
+- `CustomArticleJsonLdProps` and `CustomArticleJsonLdProps` changed
+- Meta description updated for static routes
+- Template design for social media preview
+- Google analytics only works in env variable `ANALYTICS_ENABLED`
+
 ## [2.0.1] - 2022-04-25
 ### Changed
 - Bumped `notion-compat` to `v6.12.4`